requirements.txt
pyspark==3.5.7
psycopg2-binary==2.9.10
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
prefect>=2.14.0,<3.0.0
colorama>=0.4.6